Align dates with official reporting
Until now `date` referred to the date the data was made public. Up until
early April the data reported referred to an unknown 24h period (most likely
something like 12:00-11:59); indications are that the data is now referring
to the day before i.e. 00:00-23:59.

To more accurately describe the data, all dates have now been changed
reflect the day the information applies to.

- add `date_published` which refers to the date & time the information
was made public (for timeline.csv this refers to the last update of the
day)
- add/change `date` so that it refers to the day the information applies to